Make pong increase speed with each hit

+++ firmware/src/apps/games/snake.rs
@@ -265,8 +265,8 @@ impl<'a> Game<'a> {
 
 pub async fn run() {
     let fb = Matrix::fb();
-    let mut clock_ticks = Duration::from_millis(250).as_ticks();
-    let mut clock = Ticker::every(Duration::from_ticks(clock_ticks));
+    let mut clock_ticks = Duration::from_millis(250).as_ticks() as u32;
+    let mut clock = Ticker::every(Duration::from_ticks(clock_ticks as u64));
     let mut g = Game::new(fb);
     let mut paused = false;
 
@@ -282,11 +282,11 @@ pub async fn run() {
                         break GameResult { has_won: false, score: g.score };
                     } else {
                         let new_ticks = clock_ticks / 2;
-                        if new_ticks >= matrix::FRAME_DURATION.as_ticks() {
+                        if new_ticks >= matrix::FRAME_DURATION.as_ticks() as u32 {
                             g.multiplier += 1;
 
                             clock_ticks = new_ticks;
-                            clock = Ticker::every(Duration::from_ticks(clock_ticks));
+                            clock = Ticker::every(Duration::from_ticks(clock_ticks as u64));
                         }
                     }
                 }
